feature: DeleteUserExpressionsRoute

+++ src/services/usersExpressions/DeleteUserExpressionsService.ts
@@ -5,6 +5,12 @@ export class DeleteUserExpressionsService {
 
   async execute(id: string) {
     try {
+      const expression = await this.repository.findById(id);
+      
+      if (!expression) {
+        return new Error("Expression with this ID does not exists");
+      }
+
       const result = await this.repository.delete(id);
 
       return result;
